(fix) core data sync/login/ segmented control issues
